move some widgets
