+++ packages/testsweets/lib/src/ui/widget_capture/widget_capture_widgets/inspect_layout_widget.dart
@@ -2,10 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:stacked/stacked.dart';
 import 'package:testsweets/src/constants/app_constants.dart';
 import 'package:testsweets/src/models/enums/widget_type.dart';
+import 'package:testsweets/src/ui/shared/app_colors.dart';
+import 'package:testsweets/src/ui/shared/shared_styles.dart';
 import 'package:testsweets/src/ui/widget_capture/widget_capture_viewmodel.dart';
-
-import '../app_colors.dart';
-import '../shared_styles.dart';
 
 class InspectLayoutView extends ViewModelWidget<WidgetCaptureViewModel> {
   const InspectLayoutView({
